backend/requirements.txt
dj-database-url==2.1.0
Django==4.1.7
django-awl==1.7.0
django-cors-headers==3.14.0
django-environ==0.11.2
django-react-templatetags==7.0.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
numpy==1.26.0
openai==1.13.3
pandas==1.4.4
python-dotenv==1.0.0
whitenoise==6.5.0
